Organize numba dispatches according to library
