+++ pytensor/link/numba/dispatch/__init__.py
@@ -6,13 +6,12 @@ import pytensor.link.numba.dispatch.blockwise
 import pytensor.link.numba.dispatch.compile_ops
 import pytensor.link.numba.dispatch.elemwise
 import pytensor.link.numba.dispatch.extra_ops
-import pytensor.link.numba.dispatch.nlinalg
+import pytensor.link.numba.dispatch.linalg
 import pytensor.link.numba.dispatch.random
 import pytensor.link.numba.dispatch.scan
 import pytensor.link.numba.dispatch.scalar
 import pytensor.link.numba.dispatch.shape
 import pytensor.link.numba.dispatch.signal
-import pytensor.link.numba.dispatch.slinalg
 import pytensor.link.numba.dispatch.sort
 import pytensor.link.numba.dispatch.sparse
 import pytensor.link.numba.dispatch.subtensor
