+++ pytensor/link/numba/dispatch/linalg/__init__.py
@@ -1,0 +1,5 @@
+import pytensor.link.numba.dispatch.linalg.constructors
+import pytensor.link.numba.dispatch.linalg.decomposition.dispatch
+import pytensor.link.numba.dispatch.linalg.inverse
+import pytensor.link.numba.dispatch.linalg.solve.dispatch
+import pytensor.link.numba.dispatch.linalg.summary
